lab5 - добавил проверку существования пути к файлу

+++ labs/05/src/Model/Image/Image.php
@@ -50,7 +50,9 @@ class Image implements ImageInterface
     public function __destruct()
     {
         if ($this->markDel) {
-            unlink($this->getPath());
+            if ($this->getPath() !== null) {
+                unlink($this->getPath());
+            }
         }
     }
 }
